refactor(kernel): clean up generic CPU header
